update to latest rust

+++ src/ppm/encoder.rs
@@ -87,10 +87,10 @@ fn max_pixel_value(pixel_type: color::ColorType) -> u16 {
     use std::num::Int;
 
     match pixel_type {
-        Gray(n)    => 2u16.pow(n as usize) - 1,
-        RGB(n)     => 2u16.pow(n as usize) - 1,
-        Palette(n) => 2u16.pow(n as usize) - 1,
-        GrayA(n)   => 2u16.pow(n as usize) - 1,
-        RGBA(n)    => 2u16.pow(n as usize) - 1
+        Gray(n)    => 2u16.pow(n as u32) - 1,
+        RGB(n)     => 2u16.pow(n as u32) - 1,
+        Palette(n) => 2u16.pow(n as u32) - 1,
+        GrayA(n)   => 2u16.pow(n as u32) - 1,
+        RGBA(n)    => 2u16.pow(n as u32) - 1
     }
 }
